Add files via upload
The carPrice.csv file contains data of the cars which we will use to train our model.
There is a car-price-predictor-model which is the trained model. This model is saved separately so that we don't need to train our machine every single time. We can directly use this trained model to get results. 
The Project.ipynb file contains all the steps involved to train the model and save the model.
The app.py shows how to use the saved model and host a webpage locally.

To host the webpage locally you must follow these steps:-
open cmd
use cd command to get to the directory where app.py and car-price-predictor-model is saved
then type streamlit run app.py